fix: migrate stage IDs to int64 to prevent truncation on 32-bit platforms
nextPipelineID() was casting an int64 atomic counter to int; on 32-bit
targets int is 32 bits, so at 2^31 stage IDs the cast wraps silently and
runCtx.chans map keys start colliding, producing incorrect DAG wiring.

Migrate the full ID surface to int64: stageMeta.id/inputs, Pipeline.id,
runCtx.chans key, getChan/setChan, hooks.GraphNode.ID/Inputs, and all
[]int{p.id} inputs slices in every operator file. No logic changes; pure
type propagation. Adds go.mod replace directive for local hooks development
(to be removed when hooks/v0.2.0 is tagged). Adds TestGraphNodeIDIsInt64
regression test.

+++ advanced.go
@@ -34,7 +34,7 @@ func SwitchMap[I, O any](p *Pipeline[I], fn func(context.Context, I, func(O) err
 		overflow:    cfg.overflow,
 		timeout:     cfg.timeout,
 		hasSuperv:   cfg.supervision.HasSupervision(),
-		inputs:      []int{p.id},
+		inputs:      []int64{p.id},
 	}
 	build := func(rc *runCtx) chan O {
 		if existing := rc.getChan(id); existing != nil {
@@ -174,7 +174,7 @@ func ExhaustMap[I, O any](p *Pipeline[I], fn func(context.Context, I, func(O) er
 		overflow:    cfg.overflow,
 		timeout:     cfg.timeout,
 		hasSuperv:   cfg.supervision.HasSupervision(),
-		inputs:      []int{p.id},
+		inputs:      []int64{p.id},
 	}
 	build := func(rc *runCtx) chan O {
 		if existing := rc.getChan(id); existing != nil {
@@ -300,14 +300,14 @@ func MapResult[I, O any](p *Pipeline[I], fn func(context.Context, I) (O, error),
 		kind:   "map_result",
 		name:   orDefault(cfg.name, "map_result"),
 		buffer: cfg.buffer,
-		inputs: []int{p.id},
+		inputs: []int64{p.id},
 	}
 	errMeta := stageMeta{
 		id:     errID,
 		kind:   "map_result_err",
 		name:   orDefault(cfg.name, "map_result") + "_err",
 		buffer: cfg.buffer,
-		inputs: []int{p.id},
+		inputs: []int64{p.id},
 	}
 
 	sharedBuild := func(rc *runCtx) (chan O, chan ErrItem[I]) {
@@ -384,7 +384,7 @@ func MapRecover[I, O any](p *Pipeline[I], fn func(context.Context, I) (O, error)
 		kind:   "map_recover",
 		name:   orDefault(cfg.name, "map_recover"),
 		buffer: cfg.buffer,
-		inputs: []int{p.id},
+		inputs: []int64{p.id},
 	}
 	build := func(rc *runCtx) chan O {
 		if existing := rc.getChan(id); existing != nil {
